requirements.txt
fastapi
streamlit
pydantic
numpy
dotenv
openai
uvicorn
requests
plotly
python-dateutil
motor
bcrypt
pyjwt
pymongo
boto3
pydantic[email]